// src_migration/types.ts
export interface NotificationItem {
  id?: string | number;
  tipo?: 'avisos' | 'marketing' | string;
  status?: 'nao_lida' | 'lida' | string;
  titulo?: string;
  subtitulo?: string;
  descricao?: string;
  image?: string;
  createdAt?: string;
  actions?: string;
}